moving method to shared location
