+++ programs/monaco_protocol/src/instructions/market_position/mod.rs
@@ -3,6 +3,7 @@ pub mod settle_market_position;
 pub mod update_on_order_cancellation;
 pub mod update_on_order_creation;
 pub mod update_on_order_match;
+pub mod update_product_commission_contributions;
 pub mod void_market_position;
 
 pub use create_market_position::*;
@@ -10,4 +11,5 @@ pub use settle_market_position::*;
 pub use update_on_order_cancellation::*;
 pub use update_on_order_creation::*;
 pub use update_on_order_match::*;
+pub use update_product_commission_contributions::*;
 pub use void_market_position::*;
